Major revision of folder structure
