+++ public/about.html
@@ -57,8 +57,8 @@
         <section class="content-section" id="helper-sources">
             <h2>Proudly Supported By</h2>
             <ul class="supporterlist">
-                <li><a href="https://www.sciencementorsact.org.au/"><img src="ScienceMentorsLogo.png" alt="Science Mentors ACT Logo" class="supportlogo"></a><p>Science Mentors ACT</p></li>
-                <li><a href="https://www.narrabundahc.act.edu.au/"><img src="bundahlogo.png" class="supportlogo" alt="Narrabundah College Logo"></a><p>Narrabundah College</p></li>
+                <li><a href="https://www.sciencementorsact.org.au/"><img src="assets/images/ScienceMentorsLogo.png" alt="Science Mentors ACT Logo" class="supportlogo"></a><p>Science Mentors ACT</p></li>
+                <li><a href="https://www.narrabundahc.act.edu.au/"><img src="assets/images/bundahlogo.png" class="supportlogo" alt="Narrabundah College Logo"></a><p>Narrabundah College</p></li>
             </ul>
         </section>
     </main>
