+++ public/assets/fonts/SpaceMono-Regular.ttf
