--- public/env-config.example.js
@@ -1,10 +1,0 @@
-window.PART_RUNTIME_CONFIG = {
-    PART_FIREBASE_API_KEY: "",
-    PART_FIREBASE_AUTH_DOMAIN: "",
-    PART_FIREBASE_PROJECT_ID: "",
-    PART_FIREBASE_STORAGE_BUCKET: "",
-    PART_FIREBASE_MESSAGING_SENDER_ID: "",
-    PART_FIREBASE_APP_ID: "",
-    PART_FIREBASE_MEASUREMENT_ID: "",
-    PART_ADMIN_EMAILS: ""
-};
